fix: added the cache storage in the account country selection page
