+++ packages/core/src/index.tsx
@@ -12,49 +12,9 @@ import { checkAndSetEndpointFromUrl } from '@deriv/shared';
 import AppNotificationMessages from './App/Containers/app-notification-messages.jsx';
 import { AnalyticsInitializer } from 'Utils/Analytics';
 
-let eventQueue = [];
-
-const queueEvent = event => {
-    eventQueue.push(event);
-    localStorage.setItem('pending_events', JSON.stringify(eventQueue));
-};
-
-const sendEventToAnalytics = event => {
-    console.log('Sending event:', event);
-};
-
-const loadPendingEvents = () => {
-    const storedEvents = localStorage.getItem('pending_events');
-    if (storedEvents) {
-        eventQueue = JSON.parse(storedEvents);
-    }
-};
-
-const initializeAnalytics = async () => {
-    try {
-        await AnalyticsInitializer();
-        console.log('analytics loaded');
-
-        if (eventQueue.length > 0) {
-            eventQueue.forEach(event => {
-                sendEventToAnalytics(event);
-            });
-
-            eventQueue = [];
-            localStorage.removeItem('pending_events');
-        }
-    } catch (error) {
-        console.error('Analytics failed to load', error);
-    }
-};
-
-console.log('before calling analytics');
-loadPendingEvents();
-initializeAnalytics();
-console.log('after calling analytics');
-
+AnalyticsInitializer();
 if (
-    !!window?.localStorage.getItem?.('debug_service_worker') ||
+    !!window?.localStorage.getItem?.('debug_service_worker') || // To enable local service worker related development
     (!window.location.hostname.startsWith('localhost') && !/binary\.sx/.test(window.location.hostname)) ||
     window.location.hostname === 'deriv-app.binary.sx'
 ) {
@@ -63,6 +23,7 @@ if (
 
 const has_endpoint_url = checkAndSetEndpointFromUrl();
 
+// if has endpoint url, APP will be redirected
 if (!has_endpoint_url) {
     const root_store = initStore(AppNotificationMessages);
